@@ -1,5 +1,10 @@
 package com.idormy.sms.forwarder.model;
 
+import androidx.annotation.NonNull;
+
+import lombok.Data;
+
+@Data
 public class LogModel {
     private String from;
     private String content;
@@ -14,46 +19,7 @@ public class LogModel {
         this.ruleId = ruleId;
     }
 
-    public String getFrom() {
-        return from;
-    }
-
-    public void setFrom(String from) {
-        this.from = from;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public String getSimInfo() {
-        return simInfo;
-    }
-
-    public void setSimInfo(String simInfo) {
-        this.simInfo = simInfo;
-    }
-
-    public Long getRuleId() {
-        return ruleId;
-    }
-
-    public void setRuleId(Long ruleId) {
-        this.ruleId = ruleId;
-    }
-
-    public Long getTime() {
-        return time;
-    }
-
-    public void setTime(Long time) {
-        this.time = time;
-    }
-
+    @NonNull
     @Override
     public String toString() {
         return "LogModel{" +
