@@ -1,8 +1,13 @@
 package com.idormy.sms.forwarder.model;
 
+import androidx.annotation.NonNull;
+
 import com.idormy.sms.forwarder.R;
 
+import lombok.Data;
 
+@SuppressWarnings("unused")
+@Data
 public class SenderModel {
     public static final int STATUS_ON = 1;
     public static final int STATUS_OFF = 0;
@@ -15,13 +20,12 @@ public class SenderModel {
     public static final int TYPE_SERVER_CHAN = 6;
     public static final int TYPE_TELEGRAM = 7;
     public static final int TYPE_SMS = 8;
+    public static final int TYPE_FEISHU = 9;
     private Long id;
     private String name;
     private int status;
     private int type;
-
     private String jsonSetting;
-
     private long time;
 
     public SenderModel() {
@@ -52,51 +56,12 @@ public class SenderModel {
                 return R.mipmap.serverchan;
             case (TYPE_TELEGRAM):
                 return R.mipmap.telegram;
+            case (TYPE_FEISHU):
+                return R.mipmap.feishu;
             case (TYPE_SMS):
-                return R.mipmap.sms;
             default:
                 return R.mipmap.sms;
         }
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status == STATUS_ON ? STATUS_ON : STATUS_OFF;
-    }
-
-    public int getType() {
-        return type;
-    }
-
-    public void setType(int type) {
-        this.type = type;
-    }
-
-    public String getJsonSetting() {
-        return jsonSetting;
-    }
-
-    public void setJsonSetting(String jsonSetting) {
-        this.jsonSetting = jsonSetting;
     }
 
     public int getImageId() {
@@ -117,18 +82,13 @@ public class SenderModel {
                 return R.mipmap.serverchan;
             case (TYPE_TELEGRAM):
                 return R.mipmap.telegram;
+            case (TYPE_FEISHU):
+                return R.mipmap.feishu;
+            case (TYPE_SMS):
             default:
                 return R.mipmap.sms;
 
         }
-    }
-
-    public long getTime() {
-        return time;
-    }
-
-    public void setTime(long time) {
-        this.time = time;
     }
 
     public int getSmsSimSlotId(int id) {
@@ -141,6 +101,7 @@ public class SenderModel {
         }
     }
 
+    @NonNull
     @Override
     public String toString() {
         return "SenderModel{" +
