@@ -2,51 +2,22 @@ package com.idormy.sms.forwarder.model.vo;
 
 import java.io.Serializable;
 
+import lombok.Data;
+
+@Data
 public class DingDingSettingVo implements Serializable {
     private String token;
     private String secret;
-    private String atMobils;
+    private String atMobiles;
     private Boolean atAll;
 
     public DingDingSettingVo() {
     }
 
-    public DingDingSettingVo(String token, String secret, String atMobils, Boolean atAll) {
+    public DingDingSettingVo(String token, String secret, String atMobiles, Boolean atAll) {
         this.token = token;
         this.secret = secret;
-        this.atMobils = atMobils;
-        this.atAll = atAll;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getSecret() {
-        return secret;
-    }
-
-    public void setSecret(String secret) {
-        this.secret = secret;
-    }
-
-    public String getAtMobils() {
-        return atMobils;
-    }
-
-    public void setAtMobils(String atMobils) {
-        this.atMobils = atMobils;
-    }
-
-    public Boolean getAtAll() {
-        return atAll;
-    }
-
-    public void setAtAll(Boolean atAll) {
+        this.atMobiles = atMobiles;
         this.atAll = atAll;
     }
 }
