@@ -1,24 +1,31 @@
 package com.idormy.sms.forwarder.sender;
 
-import android.os.Bundle;
 import android.os.Handler;
 import android.text.TextUtils;
 import android.util.Base64;
 import android.util.Log;
 
+import androidx.annotation.NonNull;
+
 import com.alibaba.fastjson.JSON;
 import com.idormy.sms.forwarder.utils.LogUtil;
+import com.idormy.sms.forwarder.utils.SettingUtil;
 
 import java.io.IOException;
 import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.TimeUnit;
 
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
 
+import io.reactivex.rxjava3.core.Observable;
+import io.reactivex.rxjava3.core.ObservableEmitter;
 import okhttp3.Call;
 import okhttp3.Callback;
 import okhttp3.MediaType;
@@ -27,35 +34,33 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 
-import static com.idormy.sms.forwarder.SenderActivity.NOTIFY;
+@SuppressWarnings({"ResultOfMethodCallIgnored", "rawtypes", "unchecked", "deprecation"})
+public class SenderDingdingMsg extends SenderBaseMsg {
 
-public class SenderDingdingMsg {
+    static final String TAG = "SenderDingdingMsg";
 
-    static String TAG = "SenderDingdingMsg";
-
-    public static void sendMsg(final long logId, final Handler handError, String token, String secret, String atMobiles, Boolean atAll, String msg) throws Exception {
-        Log.i(TAG, "sendMsg token:" + token + " secret:" + secret + " atMobiles:" + atMobiles + " atAll:" + atAll + " msg:" + msg);
+    public static void sendMsg(final long logId, final Handler handError, String token, String secret, String atMobiles, Boolean atAll, String content) throws Exception {
+        Log.i(TAG, "sendMsg token:" + token + " secret:" + secret + " atMobiles:" + atMobiles + " atAll:" + atAll + " content:" + content);
 
         if (token == null || token.isEmpty()) {
             return;
         }
 
-        token = "https://oapi.dingtalk.com/robot/send?access_token=" + token;
         if (secret != null && !secret.isEmpty()) {
             Long timestamp = System.currentTimeMillis();
             String stringToSign = timestamp + "\n" + secret;
             Mac mac = Mac.getInstance("HmacSHA256");
-            mac.init(new SecretKeySpec(secret.getBytes("UTF-8"), "HmacSHA256"));
-            byte[] signData = mac.doFinal(stringToSign.getBytes("UTF-8"));
+            mac.init(new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), "HmacSHA256"));
+            byte[] signData = mac.doFinal(stringToSign.getBytes(StandardCharsets.UTF_8));
             String sign = URLEncoder.encode(new String(Base64.encode(signData, Base64.NO_WRAP)), "UTF-8");
             token += "&timestamp=" + timestamp + "&sign=" + sign;
-            Log.i(TAG, "webhook_token:" + token);
+            Log.i(TAG, "token:" + token);
         }
 
         Map textMsgMap = new HashMap();
         textMsgMap.put("msgtype", "text");
         Map textText = new HashMap();
-        textText.put("content", msg);
+        textText.put("content", content);
         textMsgMap.put("text", textText);
         if (atMobiles != null || atAll != null) {
             Map AtMap = new HashMap();
@@ -83,60 +88,60 @@ public class SenderDingdingMsg {
             textMsgMap.put("at", AtMap);
         }
 
-        String textMsg = JSON.toJSONString(textMsgMap);
-        Log.i(TAG, "textMsg:" + textMsg);
+        final String requestUrl = "https://oapi.dingtalk.com/robot/send?access_token=" + token;
+        Log.i(TAG, "requestUrl:" + requestUrl);
+        final String requestMsg = JSON.toJSONString(textMsgMap);
+        Log.i(TAG, "requestMsg:" + requestMsg);
 
-        OkHttpClient client = new OkHttpClient();
-        RequestBody requestBody = RequestBody.create(MediaType.parse("application/json;charset=utf-8"),
-                textMsg);
+        Observable
+                .create((ObservableEmitter<Object> emitter) -> {
+                    Toast(handError, TAG, "开始请求接口...");
 
-        final Request request = new Request.Builder()
-                .url(token)
-                .addHeader("Content-Type", "application/json; charset=utf-8")
-                .post(requestBody)
-                .build();
-        Call call = client.newCall(request);
-        call.enqueue(new Callback() {
-            @Override
-            public void onFailure(Call call, final IOException e) {
-                LogUtil.updateLog(logId, 0, e.getMessage());
-                Log.d(TAG, "onFailure：" + e.getMessage());
+                    OkHttpClient client = new OkHttpClient();
+                    RequestBody requestBody = RequestBody.create(MediaType.parse("application/json;charset=utf-8"), requestMsg);
 
-                if (handError != null) {
-                    android.os.Message msg = new android.os.Message();
-                    msg.what = NOTIFY;
-                    Bundle bundle = new Bundle();
-                    bundle.putString("DATA", "发送失败：" + e.getMessage());
-                    msg.setData(bundle);
-                    handError.sendMessage(msg);
-                }
+                    final Request request = new Request.Builder()
+                            .url(requestUrl)
+                            .addHeader("Content-Type", "application/json; charset=utf-8")
+                            .post(requestBody)
+                            .build();
+                    Call call = client.newCall(request);
+                    call.enqueue(new Callback() {
+                        @Override
+                        public void onFailure(@NonNull Call call, @NonNull final IOException e) {
+                            LogUtil.updateLog(logId, 0, e.getMessage());
+                            Toast(handError, TAG, "发送失败：" + e.getMessage());
+                            emitter.onError(new RuntimeException("请求接口异常..."));
+                        }
 
-            }
+                        @Override
+                        public void onResponse(@NonNull Call call, @NonNull Response response) throws IOException {
+                            final String responseStr = Objects.requireNonNull(response.body()).string();
+                            Log.d(TAG, "Response：" + response.code() + "，" + responseStr);
+                            Toast(handError, TAG, "发送状态：" + responseStr);
 
-            @Override
-            public void onResponse(Call call, Response response) throws IOException {
-                final String responseStr = response.body().string();
-                Log.d(TAG, "Code：" + String.valueOf(response.code()) + responseStr);
+                            //TODO:粗略解析是否发送成功
+                            if (responseStr.contains("\"errcode\":0")) {
+                                LogUtil.updateLog(logId, 1, responseStr);
+                            } else {
+                                LogUtil.updateLog(logId, 0, responseStr);
+                            }
+                        }
+                    });
 
-                //TODO:粗略解析是否发送成功
-                if (responseStr.contains("\"errcode\":0")) {
-                    LogUtil.updateLog(logId, 1, responseStr);
-                } else {
-                    LogUtil.updateLog(logId, 0, responseStr);
-                }
-
-                if (handError != null) {
-                    android.os.Message msg = new android.os.Message();
-                    msg.what = NOTIFY;
-                    Bundle bundle = new Bundle();
-                    bundle.putString("DATA", "发送状态：" + responseStr);
-                    msg.setData(bundle);
-                    handError.sendMessage(msg);
-                    Log.d(TAG, "Coxxyyde：" + String.valueOf(response.code()) + responseStr);
-                }
-
-            }
-        });
+                }).retryWhen((Observable<Throwable> errorObservable) -> errorObservable
+                .zipWith(Observable.just(
+                        SettingUtil.getRetryDelayTime(1),
+                        SettingUtil.getRetryDelayTime(2),
+                        SettingUtil.getRetryDelayTime(3),
+                        SettingUtil.getRetryDelayTime(4),
+                        SettingUtil.getRetryDelayTime(5)
+                ), (Throwable e, Integer time) -> time)
+                .flatMap((Integer delay) -> {
+                    Toast(handError, TAG, "请求接口异常，" + delay + "秒后重试");
+                    return Observable.timer(delay, TimeUnit.SECONDS);
+                }))
+                .subscribe(System.out::println);
     }
 
 
