@@ -1,15 +1,19 @@
 package com.idormy.sms.forwarder.utils;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.util.Log;
 
+@SuppressWarnings("SynchronizeOnNonFinalField")
 public class InitUtil {
     static Boolean hasInit = false;
-    private static String TAG = "InitUtil";
+    private static final String TAG = "InitUtil";
+    @SuppressLint("StaticFieldLeak")
     private static Context context = null;
 
     public static void init(Context context1) {
-        Log.d(TAG, "TMSG init");
+        Log.d(TAG, "SmsForwarder init");
+        //noinspection SynchronizeOnNonFinalField
         synchronized (hasInit) {
             if (hasInit) return;
             hasInit = true;
